+++ R/qmap.R
@@ -31,7 +31,16 @@
 #' @export
 
 # x = NULL; y = NULL; geom = "point"; limits = NULL; shapefiles = NULL; crs = NULL; bathymetry = FALSE; glaciers = FALSE; rotate = FALSE; legends = TRUE; legend.position = "right"; lon.interval = NULL; lat.interval = NULL; ifelse(!is.null(getOption("ggOceanMaps.bathy.style")), getOption("ggOceanMaps.bathy.style"), "raster_binned_blues"); bathy.border.col = NA; bathy.size = 0.1; land.col = "grey60"; land.border.col = "black"; land.size = 0.1; gla.col = "grey95"; gla.border.col = "black"; gla.size = 0.1; grid.col = "grey70"; grid.size = 0.1; base_size = 11; projection.grid = FALSE; expand.factor = 1.1; verbose = FALSE
-qmap <- function(data, ..., x = NULL, y = NULL, geom = "point", limits = NULL, shapefiles = NULL, crs = NULL, bathymetry = FALSE, glaciers = FALSE, rotate = FALSE, legends = TRUE, legend.position = "right", lon.interval = NULL, lat.interval = NULL, bathy.style = NULL, bathy.border.col = NA, bathy.size = 0.1, land.col = "grey60", land.border.col = "black", land.size = 0.1, gla.col = "grey95", gla.border.col = "black", gla.size = 0.1, grid.col = "grey70", grid.size = 0.1, base_size = 11, projection.grid = FALSE, expand.factor = 1.1, verbose = FALSE) {
+qmap <- function(
+    data, ..., x = NULL, y = NULL, geom = "point", limits = NULL, shapefiles = NULL, 
+    crs = NULL, bathymetry = FALSE, glaciers = FALSE, rotate = FALSE, legends = TRUE, 
+    legend.position = "right", lon.interval = NULL, lat.interval = NULL, 
+    bathy.style = NULL, downsample = 0, bathy.border.col = NA, bathy.size = 0.1, 
+    bathy.alpha = 1, land.col = "grey60", land.border.col = "black", land.size = 0.1, 
+    gla.col = "grey95", gla.border.col = "black", gla.size = 0.1, grid.col = "grey70", 
+    grid.size = 0.1, base_size = 11, projection.grid = FALSE, expand.factor = 1.1, 
+    verbose = FALSE
+) {
   
   ## Coordinate columns